python 2.x to 3.x, sametime use python-2 as master

+++ primer/11/numconv.py
@@ -2,7 +2,7 @@
 def convert(func,seq):
     return[func (eachNum ) for eachNum in seq]
 
-myseq=(123,45.67,-6.2e8,99999999L)
-print convert(int,myseq)
-print convert(long,myseq)
-print convert(float,myseq)
+myseq=(123,45.67,-6.2e8,99999999)
+print(convert(int,myseq))
+print(convert(int,myseq))
+print(convert(float,myseq))
